made new class life that deals with all life information
